+++ src/extensions/index.ts
@@ -1,2 +1,4 @@
 export * from "./grid-builder/index.js";
 export * from "./weather/index.js";
+export * from "./billboard/index.js";
+export * from "./imposter/index.js";